pyproject.toml
[project]
name = "terrarepair-bench"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.34.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "scipy>=1.12.0",
    "numpy>=1.26.0",
    "rich>=13.0.0",
    "python-hcl2>=4.3.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
pythonpath = ["."]
